+++ src/bika/lims/api/__init__.py
@@ -18,6 +18,7 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+import copy
 import re
 from collections import OrderedDict
 from datetime import datetime
@@ -47,8 +48,12 @@ from plone.i18n.normalizer.interfaces import IIDNormalizer
 from plone.memoize.volatile import DontCache
 from Products.Archetypes.atapi import DisplayList
 from Products.Archetypes.BaseObject import BaseObject
+from Products.Archetypes.event import ObjectInitializedEvent
+from Products.Archetypes.utils import mapply
 from Products.CMFCore.interfaces import IFolderish
 from Products.CMFCore.interfaces import ISiteRoot
+from Products.CMFCore.permissions import ModifyPortalContent
+from Products.CMFCore.permissions import View
 from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.WorkflowCore import WorkflowException
 from Products.CMFPlone.RegistrationTool import get_member_by_login_name
@@ -153,9 +158,17 @@ def create(container, portal_type, *args, **kwargs):
     fti = types_tool.getTypeInfo(portal_type)
 
     if fti.product:
+        # create the AT object
         obj = _createObjectByType(portal_type, container, id)
-        obj.processForm()
-        obj.edit(title=title, **kwargs)
+        # update the object with values
+        edit(obj, check_permissions=False, title=title, **kwargs)
+        # auto-id if required
+        if obj._at_rename_after_creation:
+            obj._renameAfterCreation(check_auto_id=True)
+        # we are no longer under creation
+        obj.unmarkCreationFlag()
+        # notify that the object was created
+        notify(ObjectInitializedEvent(obj))
     else:
         # newstyle factory
         factory = getUtility(IFactory, fti.factory)
@@ -176,20 +189,30 @@ def create(container, portal_type, *args, **kwargs):
     return obj
 
 
-def copy_object(source, container=None, *args, **kwargs):
-    """Creates a copy of the source object into the specified container. If
-    container is None, creates the copy inside the same container as the source
+def copy_object(source, container=None, portal_type=None, *args, **kwargs):
+    """Creates a copy of the source object. If container is None, creates the
+    copy inside the same container as the source. If portal_type is specified,
+    creates a new object of this type, and copies the values from source fields
+    to the destination object. Field values sent as kwargs have priority over
+    the field values from source.
 
     :param source: object from which create a copy
     :type source: ATContentType/DexterityContentType/CatalogBrain
-    :param container: container
+    :param container: destination container
     :type container: ATContentType/DexterityContentType/CatalogBrain
+    :param portal_type: destination portal type
     :returns: The new created object
     """
+    # Prevent circular dependencies
+    from security import check_permission
+    # Use same container as source unless explicitly set
     source = get_object(source)
     if not container:
         container = get_parent(source)
-    portal_type = get_portal_type(source)
+
+    # Use same portal type as source unless explicitly set
+    if not portal_type:
+        portal_type = get_portal_type(source)
 
     # Extend the fields to skip with defaults
     skip = kwargs.pop("skip", [])
@@ -214,19 +237,73 @@ def copy_object(source, container=None, *args, **kwargs):
     skip = dict([(item, True) for item in skip])
 
     # Update kwargs with the field values to copy from source
-    for field in get_fields(source).values():
-        field_name = field.getName()
+    fields = get_fields(source)
+    for field_name, field in fields.items():
+        # Prioritize field values passed as kwargs
         if field_name in kwargs:
             continue
+        # Skip framework internal fields by name
         if skip.get(field_name, False):
             continue
-        if skip.get(field.getType(), False):
+        # Skip fields of non-suitable types
+        if hasattr(field, "getType") and skip.get(field.getType(), False):
             continue
-        field_value = field.getRaw(source)
+        # Skip readonly fields
+        if getattr(field, "readonly", False):
+            continue
+        # Skip non-readable fields
+        perm = getattr(field, "read_permission", View)
+        if perm and not check_permission(perm, source):
+            continue
+
+        # do not wake-up objects unnecessarily
+        if hasattr(field, "getRaw"):
+            field_value = field.getRaw(source)
+        elif hasattr(field, "get_raw"):
+            field_value = field.get_raw(source)
+        elif hasattr(field, "getAccessor"):
+            accessor = field.getAccessor(source)
+            field_value = accessor()
+        else:
+            field_value = field.get(source)
+
+        # Do a hard copy of value if mutable type
+        if isinstance(field_value, (list, dict, set)):
+            field_value = copy.deepcopy(field_value)
         kwargs.update({field_name: field_value})
 
     # Create a copy
     return create(container, portal_type, *args, **kwargs)
+
+
+def edit(obj, check_permissions=True, **kwargs):
+    """Updates the values of object fields with the new values passed-in
+    """
+    # Prevent circular dependencies
+    from security import check_permission
+    fields = get_fields(obj)
+    for name, value in kwargs.items():
+        field = fields.get(name, None)
+        if not field:
+            continue
+
+        # cannot update readonly fields
+        readonly = getattr(field, "readonly", False)
+        if readonly:
+            raise ValueError("Field '{}' is readonly".format(name))
+
+        # check field writable permission
+        if check_permissions:
+            perm = getattr(field, "write_permission", ModifyPortalContent)
+            if perm and not check_permission(perm, obj):
+                raise Unauthorized("Field '{}' is not writeable".format(name))
+
+        # Set the value
+        if hasattr(field, "getMutator"):
+            mutator = field.getMutator(obj)
+            mapply(mutator, value)
+        else:
+            field.set(obj, value)
 
 
 def get_tool(name, context=None, default=_marker):