+++ src/senaite/core/upgrade/v02_03_000.py
@@ -19,8 +19,6 @@
 # Some rights reserved, see README and LICENSE.
 
 from bika.lims import api
-from bika.lims.interfaces import IRejected
-from bika.lims.interfaces import IRetracted
 from Products.Archetypes.config import REFERENCE_CATALOG
 from senaite.core import logger
 from senaite.core.catalog import ANALYSIS_CATALOG
@@ -31,16 +29,13 @@ from senaite.core.catalog.report_catalog import ReportCatalog
 from senaite.core.config import PROJECTNAME as product
 from senaite.core.setuphandlers import CATALOG_MAPPINGS
 from senaite.core.setuphandlers import _run_import_step
-from senaite.core.setuphandlers import add_catalog_index
 from senaite.core.setuphandlers import add_senaite_setup
 from senaite.core.setuphandlers import setup_auditlog_catalog_mappings
 from senaite.core.setuphandlers import setup_catalog_mappings
-from senaite.core.setuphandlers import reindex_catalog_index
 from senaite.core.setuphandlers import setup_catalogs_order
 from senaite.core.setuphandlers import setup_core_catalogs
 from senaite.core.upgrade import upgradestep
 from senaite.core.upgrade.utils import UpgradeUtils
-from zope.interface import alsoProvides
 
 version = "2.3.0"  # Remember version number in metadata.xml and setup.py
 profile = "profile-{0}:default".format(product)
@@ -92,9 +87,6 @@ def upgrade(tool):
     # Add new setup folder to portal
     add_senaite_setup(portal)
 
-    # Add isAnalyte index in analyses catalog
-    add_isanalyte_index(portal)
-
     remove_stale_metadata(portal)
     fix_samples_primary(portal)
     fix_worksheets_analyses(portal)
@@ -105,7 +97,6 @@ def upgrade(tool):
     migrate_analysis_services_fields(portal)
     migrate_analyses_fields(portal)
     reindex_laboratory(portal)
-    mark_retracted_and_rejected_analyses(portal)
 
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
@@ -444,45 +435,3 @@ def reindex_laboratory(portal):
     setup = api.get_setup()
     setup.laboratory.reindexObject()
     logger.info("Reindexing laboratory content type [DONE]")
-
-
-def add_isanalyte_index(portal):
-    cat = api.get_tool(ANALYSIS_CATALOG)
-    logger.info("Add isAnalyte index to {} ...".format(cat.id))
-    if add_catalog_index(cat, "isAnalyte", "", "BooleanIndex"):
-        reindex_catalog_index(cat, "isAnalyte")
-    logger.info("Add isAnalyte index to {} [DONE]".format(cat.id))
-
-
-def mark_retracted_and_rejected_analyses(portal):
-    """Sets the IRetracted and/or IRejected interface to analyses that were
-    either retracted or rejected
-    """
-    logger.info("Applying IRetracted/IRejected interface to analyses ...")
-    query = {
-        "portal_type": ["Analysis", "ReferenceAnalysis", "DuplicateAnalysis"],
-        "review_state": ["retracted", "rejected"],
-    }
-    brains = api.search(query, ANALYSIS_CATALOG)
-    total = len(brains)
-
-    for num, brain in enumerate(brains):
-        if num and num % 100 == 0:
-            logger.info("Apply IRetracted/IRejected {0}/{1}".format(num, total))
-
-        obj = api.get_object(brain)
-        if IRetracted.providedBy(obj):
-            obj._p_deactivate()  # noqa
-            continue
-
-        if IRejected.providedBy(obj):
-            obj._p_deactivate()  # noqa
-            continue
-
-        status = api.get_review_status(obj)
-        if status == "retracted":
-            alsoProvides(obj, IRetracted)
-        elif status == "rejected":
-            alsoProvides(obj, IRejected)
-
-    logger.info("Applying IRetracted/IRejected interface to analyses [DONE]")