+++ CHANGES.rst
@@ -1,12 +1,26 @@
 Changelog
 =========
 
+2.4.0 (unreleased)
+------------------
 
-2.3.0 (unreleased)
+- #2162 Allow to create samples without analyses
+- #2160 Allow indexed attributes for ZCTextIndex in catalog API
+- #2158 Fix traceback when accessing registry
+- #2154 Cleanup the internal logic used for the creation of analysis objects
+- #2156 Fix groups for selection in login details view are hardcoded
+
+
+2.3.0 (2022-10-03)
 ------------------
 
 - #2120 Support for Multiple component analysis
 - #2150 Performance: prioritize raw getter for AllowedMethods field
+- #2153 Improve default body text for email publication to allow translations
+- #2151 Added `api.copy_object` function for both DX and AT types
+- #2150 Improve the creation process of AT content types
+- #2150 Added `api.edit` function for both DX and AT types
+- #2149 Performance: prioritize raw getter for AllowedMethods field
 - #2148 Performance: prioritize raw getter for AllowedInstruments field
 - #2147 Remove stale function workflow.getReviewHistory
 - #2146 Fix "No object found for UID: <laboratory_uid>" in report preview